+++ blueprints/flexberry-model-init/files/__root__/models/__name__.js
@@ -5,8 +5,10 @@
 %>import { buildValidations } from 'ember-cp-validations';<%
 
 if (!parentModelName) { %>
-import EmberFlexberryDataModel from 'ember-flexberry-data/models/model';
-import OfflineModelMixin from 'ember-flexberry-data/mixins/offline-model';<%
+import EmberFlexberryDataModel from 'ember-flexberry-data/models/model';<%
+}
+if (!parentModelName && additionalModelMixin && additionalModelMixinImport) { %>
+import <%= additionalModelMixin %> from <%= additionalModelMixinImport %>; <%
 } %>
 
 import {
@@ -26,7 +28,7 @@ const Validations = buildValidations(<%= parentModelName ? '$.extend({}, ParentV
   dependentKeys: ['model.i18n.locale'],
 });
 
-let Model = <%= parentModelName ? parentClassName : 'EmberFlexberryData' %>Model.extend(<%= !parentModelName ? 'OfflineModelMixin, ' : '' %><%= className %>Mixin, Validations, {
+let Model = <%= parentModelName ? parentClassName : 'EmberFlexberryData' %>Model.extend(<%= additionalModelMixin%><%= additionalModelMixin ? ', ' : '' %><%= className %>Mixin, Validations, {
 });<%
 
 if (parentModelName || projections) { %>